Use KotlinTest instead of Spek
This is entirely bikeshedding. I decided that I wanted to try out
property testing, which KotlinTest supports but Spek doesn't. The only
really unfortunate part of this is that KotlinTest doesn't play nicely
with JUnit / JUnit 5, so that and the tooling associated with it is out
for the moment. Oh well.