Move from Wasabi to Vertx
This project was origninally going to be based on Wasabi, which is a
pure-ish Kotlin web framework. However, I didn't feel that Wasabi was
giving me what I wanted in terms of, eg, binding data classes to POST
bodies. So while on a high level I would have preferred to have an
as-pure-as-possible Kotlin stack, I would rather have an application
that works.