Migrate to CircleCI 2.0
Test out a move over to CircleCI 2.0. Mostly just for fun, although in
theory I should get faster build times and I get to be more careful
about reproducible builds. In the future I also want to try out
publishing a GitHub pages site and deploying from this, so that'll be
fun.