De-emphasize Kotlin in non-product code
Unfortunately, the using Kotlin in suporting code is not perfectly
supported. In particular, even Intellij behaves very poorly with
`build.gradle.kts`, and attempting to use an unusual (or even pretty
usual) gradle plugin will plunge you into a world of hurt. In addition,
whether due to `kapt` problems or otherwise, Junit5 didn't seem to be
viewing annotations on Kotlin JUnit 5 test classes. So both the
annotation and the class are now written in Java.