Add connection to ElasticSearch
Now instead of connecting to a dummy backend the system will connect to
ElasticSearch. At least, now it has the elasic search connection in
place; it still gives dummy answers. It also now uses Dagger for
dependency injection, because I was tired of manually building the
dependency tree in the main method.