Begin adding test suite
Begin working on the test suite for the REST API part of the
application. This commit _deliberately fails_ because I am still working
on it. Believe it or not, that is a phenomenally good thing so far as I
am concerned; it means that the goddam tests run.

This also involved bumping a whole bunch of version numbers; I was
trying to get an issue where test sources weren't being generated to go
away. I think the solution ultimately was that I needed to explictly put
dagger on the `kaptTest` scope, but frankly it beats the hell out of me.