Debug ElasticSearch things
Add `BodyHandler` to Router to actually be able to access the body
of requests (why isn't that the default?) and work around
ElasticSearch's bizzare indexing structure.

You may have noticed that I am deeply disinterested in this commit;
almost all of it will get removed when I re-write this with different
tools, so it doesn't matter a heck of a lot. I'm going to push this,
cut a release, then re-write pretty much the whole application.