Move application from Vertx Web to Dropwizard
Migrate the application from a Vertx Web stack to a Dropwizard
stack. Unfortunately, for a variety of reasons this involved touching
almost every file that I created. As of this commit, the test suite
hasn't been migrated yet (and actually doesn't even compile yet).
But Dropwizard does provide exactly what I wanted from this migration,
so that is already :100: