Remove failing tests
Remove all the tests that weren't going to be working anymore. I
would have really liked to have re-written the tests to work
with Dropwizard, but unfortunately `dropwizard-testing` is coupled
to Junit 4 at the moment. I'd rather not downgrade to JUnit 5 (or
s wap to a different test framework), so resource tests are going
to get a hard pass for the moment.

It also isn't a huge deal, because the resource is as thin a wrapper
around the DAO as I could get it, and I can't run good tests of that
because ElasticSearch is a massive pain.