Swap over to a relational storage system
Move from using ElasticSearch as the backing data store to using
a relational database (in particular, H2, but a production deployment
could probably use Postgres. I'll look into it, at some point).

This involved more sweeping changes than I really wanted it to, and
unfortunately it still needs a little bit of tweaking to get it to
work perfectly.
- Biggest unexpected change was IDs from Strings to longs. Result of
  using a bigint rather than a UUID for PKs. In theory, I could
  still use a UUID, but it is neither idiomatic nor performant with
  a database. So that's life.