Clean up application configuration
Move the big block of code installing a bundle into a different area so
that the application class is less cluttered. Also install the bundle
that maps DBI errors, because that'll probably be handy.  It's possible
that it should not be installed when not in development mode, but what
do I know?