Add health check for database
Add a healthcheck for the database to Dropwizard. The built-in JDBI
healthcheck won't work here because I am using JDBI 3 for Kotlin
support, and Dropwizard works on JDBI 2 at the moment (which makes
sense, because JDBI 3 is only in beta). Anyways, the code for the
healthcheck is identical, I just needed to port it up a version and
install it manually.