Introduce parameter object
I was only passing around a `userId: String`, but I can almost guarantee
that something is going to change here. In particular, I will _probably_
start using a proper userId that joins to a user table so that I can do
auth on different keys (eg, Google, Facebook, ...). While that's for the
future, I'll need to change the data model later if I do that and I
would like to make those future changes smaller.