Attach auth to resource
Actually attach the stuff that I had prepped for auth to the resource
that needed auth protection. This was what I was aiming at this whole
time. Having said that, the auth is pretty much entirely fake at the
moment. You can say that your username is whatever you want, and it will
ignore your password entirely. So still a work in progress.