Change to multi-module build
Change the project from being a single larger module to a pair of
modules (Projects in gradle, I guess). In any case, the _ideal_ here
would be to have a new module for the front-end so I could build the
whole thing with Gradle. I think that it's unlikely that I will write
the frontend in Kotlin. I might, though!