Couple of changes:
- Bump Kotlin version to 1.1.4-3
- Add JUnit 5 dependency to :core
- Rewrite the run command in the CI file

The really interesting one is the last one. I am consistently getting
what I _think_ are OOMs without the `--info` flag on the gradle command,
but there is no reason for it that I can see. Possibly because it seems
like multiproject Kotlin builds start multiple daemons? Or maybe that's
because I don't have this globally configured so it thinks that it might
need to run different Kotlin versions for different project? Who knows.