Add configuration DSL for some Dropwizard features
A very simple wrapper around `with`, more-or-less. The advantage of
doing it this way is that in places it can be a little more stylish and
hide some nullity checks, etc. The disadvantage is that adding another
module means that the Kotlin compiler needs to start all over again,
which is _slow_. So this probably will not wind up going forward, as it
does make compilation unacceptablely slow, especially for just cleaning
up a part of the code that I rarely ever look at.