Enable a configurable log-in mechanism
I had written a psuedo auth filter that I was just using in tests.  This
makes it possible to actually start the application with the other auth
filter. This is principally for integration testing.  However, I will
likely swap the fake auth system to be Basic Auth, which still works
fine for testing and is more than sufficient for self-hosting this
application, which is something that I would like to remain a
possiblity.