Add documentation to the README
It turned out that there was a bunch more knowledge that I was carting
around for this project than I thought that there was. So I tried to get
that all down. This includes documentation for
* Testing / compiling / installing the application
* Configuring the application
* Using the application

This is really, really only a pretty trashy pass at adding
documentation, so the formatting and order is all a mess. But there's
enough stuff down that hopefully I will be able to jog my memory if I
ever drop and then return to this project.