Use Sentry for error capturing
This was something that I had planned to do for a while, but didn't plan
on doing yet. [Sentry](sentry.io) is a tool for collecting errors
(either from the logger or errors that just get thrown by the
application) and aggregating/displaying them.  This would (in theory)
make it easier for me to work with things in production (if this ever
makes it to production).