Flesh out the generating cookies / JWTs
Previously I was only very, very pretending to generate a cookie (I was
setting the cookie to `"fake-cookie"`).  Now I've swapped to sending an
actual, signed JWT and everything.  Unfortunately, it doesn't actually
read the signing key from like a file or Vault or whatever, it generates
it when the server starts up. So that's not entirely helpful.