Migrate to Drowizard 1.3
This was a pretty big release for me, because it comes with support for
Jdbi 3. This means that I don't need to write a custom healthcheck and a
bunch of other stuff! It *also* comes with support for Junit 5, but I
couldn't get the resource extension to work (it wasn't binding
parameters correctly, I think). So while that was a nice-to-have because
it would have let me remove JUnit 4 from the classpath, but oh well.