Frontend works!
Ugly as sin and has a bunch of irrelvant crap on it, but it works! There
is no way to sign in, unfortunately; You need to know the right URL to
go to, and then going there will sign you in (kinda). But it displays an
(ugly) chart! I can get the rest of this done within another week.

Big mood when you're within sight of the end of a project.