Enable HTTPS on the Dropwizard server
Previously I didn't have HTTPS turned on for this application because it
is proxied behind Nginx, which *does* have HTTPS. So the connection
between a client and the server was still secured.  Still, for a variety
of reasons that didn't work great. So this gets a secured HTTP/2 server
set up, which works better for everyone.