Use Guice for dependency injection
Rather than building up the graph by hand. I do like the ideas behind
Dagger2; but I haven't been able to get it to play well with Bazel (or
various other build systems). In any case, Guice works and is perfectly
fine for a server applications, where startup time isn't the biggest
problem in the world. So, Guice it is.