Resolve Go dep issues
It turns out that (1) Go's module system is weird, and (2) that trying
to make Go's module work and make Bazel work in the same repo is kind of
a lot of hassle. Anyways, this removes the hilarious number of
duplicated / declared but unused / otherwise broken dependencies in the
tree.  There shouldn't have ever created a problem, but certainly was
dumb and weird.