Remove Kotlin implementation of Punchcard
This has been effectively replaced by the Go version. That version has
all of the user-facing features, but doesn't have some of the other
features (eg, a verification token for OAuth & reporting to Sentry).
I'm going to need to reference back to this when I port those over, but
I don't need this code still in the repo to do that.