Consolidate on a single go.mod
Previously I was using a `go.mod` in each go source root. This played
extremely poorly with gazelle, it turns out, so I'm trying to
consolidate on one for the whole repository. I think I lose a little bit
of tooling from Go for this, which is unfortunate, but I already lost a
bunch of that by pushing for Bazel anyway.